refactor: ladder 의 기능 분리
